Change quantity type to int.

+++ src/main/java/pl/piotr/skoczylas/orderproject/dto/OrderItemDto.java
@@ -5,12 +5,12 @@ import java.math.BigDecimal;
 // w normalnym projekcie dalbym gettery i settery, a pola prywatne, ale tutaj nie bede sie w to bawil
 public class OrderItemDto {
     public BigDecimal netPrice;
-    public long quantity;
+    public int quantity;
     public BigDecimal netTotal;
     public BigDecimal total;
     public OrderDto order;
 
-    public OrderItemDto(BigDecimal netPrice, long quantity, BigDecimal netTotal, BigDecimal total) {
+    public OrderItemDto(BigDecimal netPrice, int quantity, BigDecimal netTotal, BigDecimal total) {
         this.netPrice = netPrice;
         this.quantity = quantity;
         this.netTotal = netTotal;
